requirements.txt
altgraph==0.17.5
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pywin32-ctypes==0.2.3
setuptools==80.9.0
